Move InputsForTest to macro_collector.py
This is useful to generate PSA tests for more than constant names.
